pi4/requirements.txt
face_recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
